voice-agent-server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv>=1.0.0
torch
transformers
openai-whisper
edge-tts
soundfile
numpy
google-generativeai
